feat: move config to separate file
