+++ src/components/RichTextEditor/config.ts
@@ -12,12 +12,10 @@ import {
     createLinkPlugin,
     createParagraphPlugin,
     createPlateUI,
-    createPlateUIEditor,
     createPlugins,
     createSoftBreakPlugin,
     createStrikethroughPlugin,
     createUnderlinePlugin,
-    deserializeHtml,
     ELEMENT_H1,
     ELEMENT_LINK,
     ELEMENT_PARAGRAPH,
@@ -26,15 +24,11 @@ import {
     MARK_ITALIC,
     MARK_STRIKETHROUGH,
     MARK_UNDERLINE,
-    parseHtmlDocument,
-    TDescendant,
 } from "@udecode/plate";
-import { BoldMark } from "../components/marks/bold";
-import { ItalicMark } from "../components/marks/italic";
-import { StrikethroughMark } from "../components/marks/strikethrough";
-import { UnderlineMark } from "../components/marks/underline";
-
-export const EMPTY_VALUE: TDescendant[] = [{ type: ELEMENT_PARAGRAPH, children: [{ text: "" }] }];
+import { BoldMark } from "./components/marks/bold";
+import { ItalicMark } from "./components/marks/italic";
+import { StrikethroughMark } from "./components/marks/strikethrough";
+import { UnderlineMark } from "./components/marks/underline";
 
 const components = createPlateUI({
     // this will override the components over the default ones
@@ -45,7 +39,7 @@ const components = createPlateUI({
     [MARK_STRIKETHROUGH]: StrikethroughMark,
 });
 
-const plugins = createPlugins(
+export const plugins = createPlugins(
     [
         createSoftBreakPlugin(),
         createAlignPlugin(),
@@ -71,31 +65,3 @@ const plugins = createPlugins(
         components,
     },
 );
-
-export const parseRawValue = (raw?: string): TDescendant[] => {
-    console.log({ raw });
-
-    let parsedValue = EMPTY_VALUE;
-
-    if (!raw) {
-        return parsedValue;
-    }
-
-    try {
-        parsedValue = JSON.parse(raw);
-    } catch {
-        const editor = createPlateUIEditor({ plugins: plugins });
-        const document = parseHtmlDocument(raw);
-        console.log({ document });
-        const parsedHtml = deserializeHtml(editor, {
-            element: document.body,
-            stripWhitespace: true,
-        });
-        console.log({ parsedHtml });
-        if (parsedHtml) {
-            parsedValue = parsedHtml;
-        }
-    }
-
-    return parsedValue;
-};
